Refactor #358 - migrate custom font to designSystem
