+++ core/designsystem/src/main/res/font/sb_aggro_m.ttf
